rename OriginalFrameLimiter to MainGameTime, fixed time step mode code fix
