+++ xlive/Project_Cartographer.vcxproj.filters
@@ -105,7 +105,7 @@
     <ClCompile Include="H2MOD\Modules\Networking\OverridePackets\OverridePackets.cpp" />
     <ClCompile Include="XLive\xnet\NIC.cpp" />
     <ClCompile Include="XLive\xnet\net_utils.cpp" />
-    <ClCompile Include="H2MOD\Modules\MainLoopPatches\OriginalFPSLimiter\OriginalFPSLimiter.cpp" />
+    <ClCompile Include="H2MOD\Modules\MainLoopPatches\MainGameTime\MainGameTime.cpp" />
     <ClCompile Include="Util\curl-interface.cpp" />
     <ClCompile Include="H2MOD\Modules\CustomVariantSettings\CustomVariantSettings.cpp" />
     <ClCompile Include="H2MOD\Modules\PlaylistLoader\PlaylistLoader.cpp" />
@@ -391,7 +391,7 @@
     <ClInclude Include="H2MOD\Modules\Networking\OverridePackets\OverridePackets.h" />
     <ClInclude Include="XLive\xnet\NIC.h" />
     <ClInclude Include="XLive\xnet\net_utils.h" />
-    <ClInclude Include="H2MOD\Modules\MainLoopPatches\OriginalFPSLimiter\OriginalFPSLimiter.h" />
+    <ClInclude Include="H2MOD\Modules\MainLoopPatches\MainGameTime\MainGameTime.h" />
     <ClInclude Include="Util\curl-interface.h" />
     <ClInclude Include="3rdparty\spdlog\include\spdlog\cfg\argv.h" />
     <ClInclude Include="3rdparty\spdlog\include\spdlog\cfg\env.h" />
